% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGHsim.R
\docType{data}
\name{CGHsim}
\alias{CGHsim}
\title{Simulation of CGH data}
\format{
A data frame with 400 rows and two columns:
\describe{
  \item{\code{y}}{Log R ratio}
  \item{\code{x}}{Genomic position (but in fact the row number).}
}
}
\source{
The simulation program could not be located anymore. But the data have a very simple structure.
}
\usage{
data(CGHsim)
}
\description{
A crude simulation of comparative genomic hybridization (CGH) data.
}
\keyword{datasets}
