% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_base.R
\name{clone_base}
\alias{clone_base}
\title{Clone a B-spline basis for new x}
\usage{
clone_base(B, x)
}
\arguments{
\item{B}{a B-splines basis matrix, computed with \code{bbase()} or \code{cbase()}.}

\item{x}{a vector of new argument values.}
}
\value{
A matrix with number of rows=\code{length(xnew)}.
}
\description{
Extract basis parameters from an existing B-splines basis matrix,
and use them for computing a new basis at new values of \code{x}.
}
\details{
If values in \code{x} are outside the domain used for computing \code{B}, they will be discarded, with a warning.
}
\examples{

x = seq(0, 10, length = 20)
n = length(x)
y = sin(x / 2) + rnorm(n) * 0.2
B = bbase(x)
nb = ncol(B)
D = diff(diag(nb), diff = 2)
lambda = 1
a = solve(t(B) \%*\% B + lambda * t(D)\%*\% D, t(B) \%*\% y)
# Clone basis on finer grid
xg = seq(0, 10, length = 200)
Bg = clone_base(B, xg)
yg = Bg \%*\% a
plot(x, y)
lines(xg, yg, col = 'blue')

}
\references{
Eilers, P.H.C. and Marx, B.D. (2021). \emph{Practical Smoothing, The Joys of
P-splines.} Cambridge University Press.

Eilers, P.H.C., Marx, B.D., and Durban, M.
(2015). Twenty years of P-splines, \emph{SORT}, 39(2): 149-186.
}
\author{
Paul Eilers
}
