% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G519C18.R
\docType{data}
\name{G519C18}
\alias{G519C18}
\title{Chromosome G519C18 data}
\format{
A dataframe with two columns:
\describe{
  \item{\code{y}}{Probe position}
  \item{\code{x}}{Log R Ratio}. }
}
\usage{
data(G519C18)
}
\description{
An extract of the data set \code{G519} in the Bioconductor package \code{Vega}, for chromosome 18.
}
\examples{
plot(G519C18$x, G519C18$y, type = 'l', ylab = 'LRR', xlab = 'Position', main = 'Chromosome 18')
}
\references{
https://www.bioconductor.org/packages/release/bioc/html/Vega.html
}
\keyword{datasets}
