% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greece_deaths.R
\docType{data}
\name{Greece_deaths}
\alias{Greece_deaths}
\title{Deaths in Greece in 1960.}
\format{
A dataframe with three columns:
\describe{
  \item{\code{Age}}{0 - 85}
  \item{\code{Male}}{male deaths}
  \item{\code{Female}}{female deaths.}
}
}
\source{
Personal information from Aris Perperoglou.
}
\usage{
data(Greece_deaths)
}
\description{
{Deaths in Greece in 1960.}
}
\details{
All counts for ages above 84 have been grouped to one number for age 85.
}
\keyword{datasets}
