% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hepatitis.R
\docType{data}
\name{Hepatitis}
\alias{Hepatitis}
\title{Prevalence of Hepatitis among a sample of Bulgarian males.}
\format{
A data frame with three columns:
\describe{
\item{\code{Age}}{years}
\item{\code{Infected}}{number of infected persons}
\item{\code{Sampled}}{number of sampled persons.}
}
}
\source{
Table 2 in Keiding (1991).
}
\usage{
data(Hepatitis)
}
\description{
Prevalence of Hepatitis among a sample of Bulgarian males.
}
\references{
N. Keiding (1991) Age-Specific Incidence and Prevalence: A Statistical Perspective.
\emph{JRSS-A} 154, 371-396.
}
\keyword{datasets}
