% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_link.R
\name{inverse_link}
\alias{inverse_link}
\title{Inverse link function, used for GLM fitting.}
\usage{
inverse_link(x, link)
}
\arguments{
\item{x}{scalar, vector, or matrix input.}

\item{link}{the link function, one of \code{"identity"}, \code{"log"}, \code{"sqrt"},
\code{"logit"}, \code{"probit"}, \code{"cloglog"}, \code{"loglog"}, \code{"reciprocal"};
quotes are needed (default \code{"identity"}).}
}
\value{
The inverse link function applied to \code{x}.
If \code{link} is not in the above list of allowed  names, \code{NULL} will be returned.
}
\description{
Inverse link function, used for GLM fitting.
}
