% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOPS_helpers.R
\name{JOPS_colors}
\alias{JOPS_colors}
\title{Custom color ramp.}
\usage{
JOPS_colors(n)
}
\arguments{
\item{n}{number of steps.}
}
\value{
custom color ramp.
}
\description{
Custom color ramp.
}
\references{
Eilers, P.H.C. and Marx, B.D. (2021). \emph{Practical Smoothing, The Joys of
P-splines.} Cambridge University Press.
}
