% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOPS_helpers.R
\name{JOPS_point}
\alias{JOPS_point}
\title{Themeing functions used to unify ggplot features}
\usage{
JOPS_point(s_size = 1.5)
}
\arguments{
\item{s_size}{point size parameter for ggplot2 (default = 1.5).}
}
\value{
themeing function for ggplot2 features.
}
\description{
Custom size and color of points.
}
