% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JOPS_helpers.R, R/JOPS_theme.R
\name{JOPS_theme}
\alias{JOPS_theme}
\title{Custom theme for ggplot}
\usage{
JOPS_theme(h_just = 0.5)

JOPS_theme(h_just = 0.5)
}
\arguments{
\item{h_just}{horizontal justification for ggplot2.}
}
\value{
custom theme for ggplot.

Custom theming function used to unify ggplot features.
}
\description{
Set a ggplot theme in black and white, with centered titles.

Set a ggplot theme in black and white, with centered titles.
}
\references{
Eilers, P.H.C. and Marx, B.D. (2021). \emph{Practical Smoothing, The Joys of
P-splines.} Cambridge University Press.

Eilers, P.H.C. and Marx, B.D. (2021). \emph{Practical Smoothing, The Joys of
P-splines.} Cambridge University Press.
}
