% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ova.R
\docType{data}
\name{ova}
\alias{ova}
\title{Ovarian cancer data}
\format{
A dataframe with five columns:

\describe{
  \item{\code{Diameter}}{}
  \item{\code{FIGO}}{}
  \item{\code{Karnofsky}}{}
  \item{\code{time}}{}
  \item{\code{death}}{death}
}
}
\source{
tba
}
\usage{
data(ova)
}
\description{
Ovarian cancer data
}
\keyword{datasets}
