% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_PDF.R
\name{save_PDF}
\alias{save_PDF}
\title{Save a plot as a PDF file.}
\usage{
save_PDF(
  fname = "scratch",
  folder = "../../Graphs",
  show = T,
  width = 6,
  height = 4.5
)
}
\arguments{
\item{fname}{the file name without the extension PDF (default: \code{scratch}).}

\item{folder}{the folder for saving PDF plots (default \code{../../Graphs)}.}

\item{show}{a logical parameter; if TRUE the full file name will be displayed.}

\item{width}{figure width in inches (default = 6).}

\item{height}{figure height in inches (default = 4.5).}
}
\value{
save a plot as a PDF file.
}
\description{
Save a plot as a PDF file in a (default) folder.
The present default is determined by the folder structure for the production of the book.
}
\references{
Eilers, P.H.C. and Marx, B.D. (2021). \emph{Practical Smoothing, The Joys of
P-splines.} Cambridge University Press.
}
\author{
Paul Eilers
}
