% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_panels.R
\name{set_panels}
\alias{set_panels}
\title{Prepare graphics layout for multiple panels}
\usage{
set_panels(rows = 1, cols = 1)
}
\arguments{
\item{rows}{number of rows.}

\item{cols}{number of columns.}
}
\value{
Prepare graphics layout for multiple panels
}
\description{
Adapt margins and axes layout for multiple panels.
}
\references{
Eilers, P.H.C. and Marx, B.D. (2021). \emph{Practical Smoothing, The Joys of
P-splines.} Cambridge University Press.
}
\author{
Paul Eilers
}
