% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_window.R
\name{set_window}
\alias{set_window}
\title{Open a graphics window.}
\usage{
set_window(width = 6, height = 4.5, kill = TRUE, noRStudioGD = TRUE)
}
\arguments{
\item{width}{figure width in inches (default = 6).}

\item{height}{figure height in inches (default = 4.5).}

\item{kill}{if TRUE (default) closes all graphics windows. Works only for Windows.}

\item{noRStudioGD}{if TRUE: do not use the RStudio device (which does not accept width and height).}
}
\value{
open a graphics window.
}
\description{
Open a a window for graphics, with specified width and height.
}
\note{
Currently only works for Windows!
}
\references{
Eilers, P.H.C. and Marx, B.D. (2021). \emph{Practical Smoothing, The Joys of
P-splines.} Cambridge University Press.
}
