% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suicide.R
\docType{data}
\name{Suicide}
\alias{Suicide}
\title{Suicide Data Set}
\format{
A dataframe with one column: \code{y}.
}
\source{
Silverman, B. (1986).
Density Estimation for Statistics and Data Analysis. Chapman & Hall.
}
\usage{
data(Suicide)
}
\description{
The dataset comprises lengths (in days) of psychiatric treatment spells for patients used as controls
in a study of suicide risks.
}
\references{
Silverman, B. (1986).
Density Estimation for Statistics and Data Analysis. Chapman & Hall.
}
\keyword{datasets}
