% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woodsurf.R
\docType{data}
\name{Woodsurf}
\alias{Woodsurf}
\title{Profile of a sanded piece of wood.}
\format{
A data frame with one column: \code{y}.
}
\source{
Pandit, S.M. and Wu, S.M. (1993).
\emph{Time Series and System Analysis with Applications}.
Krieger Publishing Company.
}
\usage{
data(Woodsurf)
}
\description{
Profile of a sanded piece of wood.
}
\keyword{datasets}
